refactor: move cds extraction into standalone file
