+++ packages_rs/nextclade/src/translate/mod.rs
@@ -1,5 +1,6 @@
 pub mod aa_alignment_ranges;
 pub mod complement;
+pub mod extract;
 pub mod frame_shifts_detect;
 pub mod frame_shifts_flatten;
 pub mod frame_shifts_translate;
